requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
django-cors-headers
pycryptodome